Had difficulty getting this to work as spring security caused an issue.
It has been a good few weeks since my last interaction with it, fixed by adding WithMockUser and with(csrf())

Also had to drop from an exact match to a loose match with any(User.class), although I could have fixed it by redefining a user object. It is late though so I will not be doing that.

Added spring security test to use extra annotation and csrf